// Fontend/TNT.N8.Web/TNT.N8.Web/ClientApp/src/app/shared/models/cauHinhPhanHangKh.model.ts
export class CauHinhPhanHangKhModel {
    id: string | null;
    parentId: string | null;
    dieuKienId: number | null;
    dieuKien: string;
    giaTriTu: number | null;
    giaTriDen: number | null;
    phanHangId: string | null;
    phanHang: string;
    createdDate: string | null;
    createdById: string | null;
    updatedDate: string | null;
    updatedById: string | null;
    tenantId: string | null;
}